Add test case to prevent missing cases
The quad files are now only written if the test case passes.
The test case generates random expressions with four numbers and
tests that there is an expression tree generated by `genExprs` which
evaluates to the same number as that of the randomly generated expression.
I hope this makes up for the silly times I missed some expressions
earlier.